requirements.txt
py-cpuinfo